Refactored Profile page to have UserProfile and ClientProfile components and render based on user; Incorporated ClientProfileInfo into ClientProfile
